Update code to match changes in alligator interfaces.

+++ apis/alligator/src/fr/lip6/move/coloane/api/alligator/dialog/FloatDialog.java
@@ -42,7 +42,7 @@ public final class FloatDialog
     public
         void updateParameter() {
         if (_errorMessage() == null) {
-            parameter.setValue(Float.valueOf(input.getText()));
+            parameter.setValue(Double.valueOf(input.getText()));
             page.refresh();
         }
     }
